apps/api/v1/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
alembic
asyncpg
pydantic-settings
email-validator
passlib[bcrypt]
python-jose[cryptography]
httpx
python-jose[cryptography]